faceandeyedetect.py
Introduction:-
Face detection is a computer vision technology that helps to
locate/visualize human faces in digital images. This technique is a
specific use case of object detection technology that deals with
detecting instances of semantic objects of a certain class (such as
humans, buildings or cars) in digital images and videos. With the
advent of technology, face detection has gained a lot of importance
especially in fields like photography, security, and marketing.

Package Used:
 CV2 Library (Computer vision)
OpenCV-Python is a library of Python bindings designed to solve computer vision
problems.
Python is a general purpose programming language started by Guido van
Rossum that became very popular very quickly, mainly because of its simplicity
and code readability. It enables the programmer to express ideas in fewer lines
of code without reducing readability.
Compared to languages like C/C++, Python is slower. That said, Python can be
easily extended with C/C++, which allows us to write computationally intensive
code in C/C++ and create Python wrappers that can be used as Python modules.
This gives us two advantages: first, the code is as fast as the original C/C++
code (since it is the actual C++ code working in background) and second, it
easier to code in Python than C/C++. OpenCV-Python is a Python wrapper for
the original OpenCV C++ implementation.